requirements.txt
requests==2.32.3
feedparser==6.0.11
openai==1.77.0
PyYAML==6.0.2
python-dotenv==1.1.0
pytz==2025.1
